Uncouple Collections from Illuminate\Support namespace
