+++ src/Illuminate/Console/Scheduling/Schedule.php
@@ -134,7 +134,7 @@ class Schedule
      * Get all of the events on the schedule that are due.
      *
      * @param  \Illuminate\Contracts\Foundation\Application  $app
-     * @return \Illuminate\Support\Collection
+     * @return \Illuminate\Collection\Collection
      */
     public function dueEvents($app)
     {
