+++ src/Illuminate/Contracts/Notifications/Dispatcher.php
@@ -7,7 +7,7 @@ interface Dispatcher
     /**
      * Send the given notification to the given notifiable entities.
      *
-     * @param  \Illuminate\Support\Collection|array|mixed  $notifiables
+     * @param  \Illuminate\Collection\Collection|array|mixed  $notifiables
      * @param  mixed  $notification
      * @return void
      */
@@ -16,7 +16,7 @@ interface Dispatcher
     /**
      * Send the given notification immediately.
      *
-     * @param  \Illuminate\Support\Collection|array|mixed  $notifiables
+     * @param  \Illuminate\Collection\Collection|array|mixed  $notifiables
      * @param  mixed  $notification
      * @return void
      */
