+++ src/Illuminate/Database/Concerns/BuildsQueries.php
@@ -129,7 +129,7 @@ trait BuildsQueries
     /**
      * Create a new length-aware paginator instance.
      *
-     * @param  \Illuminate\Support\Collection  $items
+     * @param  \Illuminate\Collection\Collection  $items
      * @param  int  $total
      * @param  int  $perPage
      * @param  int  $currentPage
@@ -146,7 +146,7 @@ trait BuildsQueries
     /**
      * Create a new simple paginator instance.
      *
-     * @param  \Illuminate\Support\Collection  $items
+     * @param  \Illuminate\Collection\Collection  $items
      * @param  int $perPage
      * @param  int $currentPage
      * @param  array  $options
