+++ src/Illuminate/Database/Eloquent/Builder.php
@@ -668,7 +668,7 @@ class Builder
      *
      * @param  string  $column
      * @param  string|null  $key
-     * @return \Illuminate\Support\Collection
+     * @return \Illuminate\Collection\Collection
      */
     public function pluck($column, $key = null)
     {
