+++ src/Illuminate/Http/Resources/Json/ResourceResponse.php
@@ -2,7 +2,7 @@
 
 namespace Illuminate\Http\Resources\Json;
 
-use Illuminate\Support\Collection;
+use Illuminate\Collection\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Contracts\Support\Responsable;
