+++ src/Illuminate/Http/Resources/MergeValue.php
@@ -2,7 +2,7 @@
 
 namespace Illuminate\Http\Resources;
 
-use Illuminate\Support\Collection;
+use Illuminate\Collection\Collection;
 
 class MergeValue
 {
@@ -16,7 +16,7 @@ class MergeValue
     /**
      * Create new merge value instance.
      *
-     * @param  \Illuminate\Support\Collection|array  $data
+     * @param  \Illuminate\Collection\Collection|array  $data
      * @return void
      */
     public function __construct($data)
