+++ src/Illuminate/Mail/Mailable.php
@@ -6,8 +6,8 @@ use ReflectionClass;
 use ReflectionProperty;
 use BadMethodCallException;
 use Illuminate\Support\Str;
-use Illuminate\Support\Collection;
 use Illuminate\Container\Container;
+use Illuminate\Collection\Collection;
 use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Contracts\Queue\Factory as Queue;
 use Illuminate\Contracts\Mail\Mailer as MailerContract;
