+++ src/Illuminate/Pagination/LengthAwarePaginator.php
@@ -6,8 +6,8 @@ use Countable;
 use ArrayAccess;
 use JsonSerializable;
 use IteratorAggregate;
-use Illuminate\Support\Collection;
 use Illuminate\Support\HtmlString;
+use Illuminate\Collection\Collection;
 use Illuminate\Contracts\Support\Jsonable;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator as LengthAwarePaginatorContract;
