+++ src/Illuminate/Support/Testing/Fakes/NotificationFake.php
@@ -3,7 +3,7 @@
 namespace Illuminate\Support\Testing\Fakes;
 
 use Ramsey\Uuid\Uuid;
-use Illuminate\Support\Collection;
+use Illuminate\Collection\Collection;
 use PHPUnit\Framework\Assert as PHPUnit;
 use Illuminate\Contracts\Notifications\Factory as NotificationFactory;
 
@@ -100,7 +100,7 @@ class NotificationFake implements NotificationFactory
      * @param  mixed  $notifiable
      * @param  string  $notification
      * @param  callable|null  $callback
-     * @return \Illuminate\Support\Collection
+     * @return \Illuminate\Collection\Collection
      */
     public function sent($notifiable, $notification, $callback = null)
     {
@@ -150,7 +150,7 @@ class NotificationFake implements NotificationFactory
     /**
      * Send the given notification to the given notifiable entities.
      *
-     * @param  \Illuminate\Support\Collection|array|mixed  $notifiables
+     * @param  \Illuminate\Collection\Collection|array|mixed  $notifiables
      * @param  mixed  $notification
      * @return void
      */
@@ -162,7 +162,7 @@ class NotificationFake implements NotificationFactory
     /**
      * Send the given notification immediately.
      *
-     * @param  \Illuminate\Support\Collection|array|mixed  $notifiables
+     * @param  \Illuminate\Collection\Collection|array|mixed  $notifiables
      * @param  mixed  $notification
      * @return void
      */
