+++ tests/Database/DatabaseEloquentCollectionTest.php
@@ -5,7 +5,7 @@ namespace Illuminate\Tests\Database;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
 use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Support\Collection as BaseCollection;
+use Illuminate\Collection\Collection as BaseCollection;
 
 class DatabaseEloquentCollectionTest extends TestCase
 {
